+++ Runtime/Module/Advertising/IAffiseAdvertisingApi.cs
@@ -1,0 +1,10 @@
+﻿#nullable enable
+using AffiseAttributionLib.Modules;
+
+namespace AffiseAttributionLib.Module.Advertising
+{
+    public interface IAffiseAdvertisingApi : IAffiseModuleApi
+    {
+        void StartModule();
+    }
+}