+++ Runtime/Module/Advertising/IAffiseModuleAdvertisingApi.cs
@@ -1,0 +1,8 @@
+﻿#nullable enable
+namespace AffiseAttributionLib.Module.Advertising
+{
+    public interface IAffiseModuleAdvertisingApi : IAffiseAdvertisingApi, IAffiseHasModule
+    {
+        
+    }
+}